[#492] Fix clang-tidy: modernize-deprecated-headers

+++ test/language/expressions/cpp/NumBitsOperatorTest.cpp
@@ -1,4 +1,4 @@
-#include "math.h"
+#include <cmath>
 
 #include "gtest/gtest.h"
 
@@ -19,7 +19,7 @@ namespace
         if (value == 1)
             return 1;
 
-        return static_cast<uint8_t>(log2(value - 1) + 1);
+        return static_cast<uint8_t>(log2(static_cast<double>(value - 1)) + 1);
     }
 }
